Génère un rapport en HTML à partir d'un fichier Quarto

Cette fonction génère un rapport en HTML à partir d'un fichier Quarto spécifié en utilisant les paramètres de la commune et du département. Le rapport est rendu en HTML et enregistré à l'emplacement spécifié par l'utilisateur.

Usage

generer_rapport(commune, departement, output)

Arguments

commune

Code de la commune. Doit être un nombre ou une chaîne de caractères représentant le code de la commune.

departement

Code du département. Doit être un nombre ou une chaîne de caractères représentant le code du département.

output

Chemin complet où le fichier HTML généré sera sauvegardé.

Value

Aucun retour, mais un fichier HTML est généré et sauvegardé à l'emplacement spécifié.

Examples

# Génère un rapport pour la commune avec code 44109 et le département avec code 32 avec un chemin de sortie personnalisé
generer_rapport(44109, 32, "output/mon_rapport.html")
#> ERROR: --output option cannot specify a relative or absolute path
#> 
#> Stack trace:
#>     at parseRenderFlags (file:///C:/PROGRA~1/RStudio/RESOUR~1/app/bin/quarto/bin/quarto.js:20804:31)
#>     at Command.actionHandler (file:///C:/PROGRA~1/RStudio/RESOUR~1/app/bin/quarto/bin/quarto.js:83068:25)
#>     at Command.execute (file:///C:/PROGRA~1/RStudio/RESOUR~1/app/bin/quarto/bin/quarto.js:8017:24)
#>     at eventLoopTick (ext:core/01_core.js:153:7)
#>     at async Command.parseCommand (file:///C:/PROGRA~1/RStudio/RESOUR~1/app/bin/quarto/bin/quarto.js:7907:20)
#>     at async quarto (file:///C:/PROGRA~1/RStudio/RESOUR~1/app/bin/quarto/bin/quarto.js:118224:9)
#>     at async file:///C:/PROGRA~1/RStudio/RESOUR~1/app/bin/quarto/bin/quarto.js:118244:9
#>     at async mainRunner (file:///C:/PROGRA~1/RStudio/RESOUR~1/app/bin/quarto/bin/quarto.js:118128:9)
#>     at async file:///C:/PROGRA~1/RStudio/RESOUR~1/app/bin/quarto/bin/quarto.js:118235:5
#> 
#> Error in quarto::quarto_render(input = "inst/rapport.qmd", output_format = "html",     output_file = output, execute_params = list(code_commune = commune,         code_departement = departement)): ✖ Error running quarto cli.
#> Caused by error:
#> ! System command 'quarto.exe' failed